namespace finish_lines, add tests

+++ app/controllers/places/finish_lines_controller.rb
@@ -4,14 +4,19 @@ module Places
     before_action :load_finish_line, only: %i[show edit update destroy]
 
     def index
+      authorize Place::FinishLine
       @finish_lines = @place.finish_lines
     end
 
     def new
+      authorize Place::FinishLine
+
       @finish_line = @place.finish_lines.new
     end
 
     def create
+      authorize Place::FinishLine
+
       finish_line = @place.finish_lines.new(finish_line_params)
 
       respond_to do |format|
@@ -24,13 +29,17 @@ module Places
     end
 
     def show
+      authorize @finish_line
       fresh_when etags_for(@finish_line)
     end
 
     def edit
+      authorize @finish_line
     end
 
     def update
+      authorize @finish_line
+
       respond_to do |format|
         if @finish_line.update(finish_line_params)
           format.js { redirect_to place_finish_lines_path(@place) }
@@ -41,6 +50,8 @@ module Places
     end
 
     def destroy
+      authorize @finish_line
+
       respond_to do |format|
         if @finish_line.destroy
           format.js { redirect_to place_finish_lines_path(@place) }
@@ -61,7 +72,7 @@ module Places
     end
 
     def finish_line_params
-      params.require(:finish_line).permit(
+      params.require(:place_finish_line).permit(
         :name,
         :start_latitude,
         :start_longitude,
